fixes for linux support

+++ drunk_snail/c/compile/approaches/append/compileAppend.c
@@ -7,7 +7,7 @@ void compileAppend_(
 	char *template_name,
 	Tree *templates_tree,
 	char **buf,
-	__int64 buffer_size,
+	int buffer_size,
 	int inner_tabs_number,
 	char *prefix_start,
 	char *prefix_end,
@@ -94,7 +94,7 @@ static PyObject *compileAppend (
 ) {
 
 	char *name;
-	__int64 buffer_size;
+	int buffer_size;
 	
 	if (!PyArg_ParseTuple(args, "si", &name, &buffer_size)) {
 		return PyLong_FromLong(-1);
