+++ drunk_snail/c/compile/approaches/comprehension/compileComprehension.c
@@ -7,7 +7,7 @@ void compileComprehension_(
 	char *template_name,
 	Tree *templates_tree,
 	char **buf,
-	__int64 buffer_size,
+	int buffer_size,
 	int inner_tabs_number,
 	char *prefix_start,
 	char *prefix_end,
@@ -98,7 +98,7 @@ static PyObject *compileComprehension (
 ) {
 
 	char *name;
-	__int64 buffer_size;
+	int buffer_size;
 	
 	if (!PyArg_ParseTuple(args, "si", &name, &buffer_size)) {
 		return PyLong_FromLong(-1);
